requirements.txt
streamlit
requests
docker
python-dotenv
